Make `split_varnames` method private; improve performance for DynamicPPL functions (#85)
Closes #84.

```julia
using Turing, Random, LinearAlgebra
y = [28, 8, -3, 7, -1, 1, 18, 12]
sigma = [15, 10, 16, 11, 9, 11, 10, 18]
@model function eight_schools(y, sigma)
    mu ~ Normal(0, 5)
    tau ~ truncated(Cauchy(0, 5); lower=0)
    theta ~ MvNormal(fill(mu, length(y)), tau^2 * I)
    for i in eachindex(y)
        y[i] ~ Normal(theta[i], sigma[i])
    end
    return (mu=mu, tau=tau)
end
model = eight_schools(y, sigma)

using FlexiChains
chn = sample(Xoshiro(468), model, NUTS(), 1000; progress=false, chain_type=VNChain)
@time returned(model, chn);
@time returned(model, chn);

# main
# first run: 0.949556 seconds (4.62 M allocations: 224.590 MiB, 3.21% gc time, 93.83% compilation time)
# subsequent: 0.061431 seconds (750.36 k allocations: 31.561 MiB)

# This PR
# first run: 0.626965 seconds (2.62 M allocations: 135.214 MiB, 2.50% gc time, 98.09% compilation time)
# subsequent: 0.014705 seconds (169.83 k allocations: 13.120 MiB)

# For comparison: MCMCChains
# first run: 0.905221 seconds (4.94 M allocations: 255.595 MiB, 3.71% gc time, 89.77% compilation time)
# subsequent: 0.074779 seconds (394.27 k allocations: 17.673 MiB)
```

The cost of this PR is that, since we no longer use `InitFromParams`
which indirectly uses `AbstractPPL.getvalue(params, varname, dist)`, we
sacrifice the ability to run `returned`, `predict`, etc. on chains where
VarNames have been split up. That is to say, this:

```julia
returned(model, FlexiChains.split_varnames(chn))
```

used to work exactly the same as `returned(model, chn)`; but now it
won't (unless of course `split_varnames(chn) == chn`).

For `returned`, it will error loudly, but for `predict` it will lead to
incorrect results since the un-split VarNames will be sampled anew from
the prior.

For this reason, this PR also makes `split_varnames` a private function.
It is still used in summaries and plotting, but apart from that it is
not offered as an option to the user.

Obviously, this is only a problem if the user manually calls
`split_varnames`. If they don't, then all is well.

Consequently, this is a breaking change.